requirements.txt
selenium==3.141.0
selene==2.0.0a38
curlify==2.2.1
pytest==6.0.1
pytest-xdist==2.0.0
pytest-rerunfailures==9.0
pytest-lazy-fixture==0.6.3
allure-pytest==2.8.18
urllib3== 1.25.10
requests==2.24.0
PySocks==1.7.1
Faker==4.1.1
pytz==2020.1
waiting==1.4.1
assertpy==1.1
pdfminer.six==20200517
deepdiff==5.0.2
dacite==1.5.1
python-dateutil==2.8.1
jsonschema==3.2.0
